# lib/api-zod/src/generated/types/challengeSubmission.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * WardSpace non-clinical prototype API
 * OpenAPI spec version: 0.1.0
 */
import type { ChallengeSubmissionStatus } from './challengeSubmissionStatus';

export interface ChallengeSubmission {
  id: number;
  challengeId: number;
  /** @maxLength 1000 */
  text: string;
  status: ChallengeSubmissionStatus;
  createdAt: string;
}