# lib/api-zod/src/generated/types/challengeSubmissionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * WardSpace non-clinical prototype API
 * OpenAPI spec version: 0.1.0
 */

export interface ChallengeSubmissionInput {
  /**
     * @minLength 1
     * @maxLength 1000
     */
  text: string;
}