# lib/api-zod/src/generated/types/challengeSubmissionReceipt.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * WardSpace non-clinical prototype API
 * OpenAPI spec version: 0.1.0
 */

export interface ChallengeSubmissionReceipt {
  received: boolean;
}